[rest] api implementation

+++ rest/scripts/task-worker.js
@@ -1,11 +1,23 @@
+const path = require('path');
+const fs = require('fs');
 const Task = require('./task');
 const Utils  = require('./utils');
 
-const taskProperties = Task.getRenderedPropertiesNamesAsList();
+const taskProperties = Task.getPropertiesNamesAsList();
 const NEW_ITEM_INDEX = Task.getNewItemIndex();
 
-const path = require('path');
-const fs = require('fs');
+const statuses = {
+    successNoContent: 204,
+    notFound: 404,
+    unprocessableEntity: 422,
+};
+
+//
+// code that uses this class should call updateJsonStorage()
+// todo
+//  implement router that calls updateJsonStorage()
+//  ? implement router that updates attachment storage ?
+//
 
 class TaskWorker {
 
@@ -13,6 +25,10 @@ class TaskWorker {
         this.tasks = [];
         this.initializeJsonStorage();
         this.initializeAttachmentsStorage()
+    }
+
+    getImplementedStatuses(){
+        return {...statuses};
     }
 
     getTasksDirectory() {
@@ -57,57 +73,109 @@ class TaskWorker {
         }
     }
 
-    getTasksCount() {
+    isItemExists(id) {
+        return Boolean(this.tasks[id]);
+    }
+    getNewItemIndex(){
         return (this.tasks === null)
             ? 0
             : this.tasks.length;
     }
-    isIndexValid(id) {
-        return !((id > this.getTasksCount() || id < 0) && (id !== NEW_ITEM_INDEX));
-    }
-    getNewItemIndex(){
-        return this.getTasksCount();
-    }
 
-    getTasksAsRenderedData() {
+    getAttachmentPathById(id){
+        if (this.tasks[id]) {
+            return this.tasks[id].getAttachmentPath();
+        }
+        return null;
+    }
+    getTasksData() {
         const tasks = [];
         for (let value of this.tasks) {
             if (value) {
-                tasks.push(value.getRenderedData());
+                tasks.push(value.getData());
             }
         }
         return tasks;
     }
+    getTaskDataById(id){
+        if (this.tasks[id]) {
+            return this.tasks[id].getData();
+        }
+        return null;
+    }
 
     completeTask(id) {
+
+        if (!this.isItemExists(id)) {
+            return statuses.notFound;
+        }
+
         this.tasks[id].changeCompleteness(true);
+
+        return statuses.successNoContent;
     }
 
     //
-    // deleteTask() / changeTask() / createTask() affects filesystem
+    // deleteTask() / changeTask() / createTask() affect filesystem
     //
 
     deleteTask(id) {
+        if (!this.isItemExists(id)) {
+            return statuses.notFound;
+        }
+
         this.tasks[id] = null;
         Utils.deleteFolderWithAttachment(path.join(this.getAttachmentsDirectory(), String(id)));
+        return statuses.successNoContent;
     }
-    changeTask(properties, attachment, rewriteAttachment, taskId) {
+
+    // properties is { },
+    // following properties of this object are defined:
+    // properties[taskProperties.taskDate]
+    // properties[taskProperties.taskCompleted]
+    // properties[taskProperties.taskName]
+    // etc
+
+    changeTask(properties, attachment, taskId) {
+
+        if (!this.isItemExists(taskId)) {
+           return statuses.notFound;
+        }
+        if (!Task.isValidObject(properties, false)) {
+            return statuses.unprocessableEntity;
+        }
+
         const task = this.tasks[taskId];
         task.changeDate(new Date(properties[taskProperties.taskDate]));
         task.changeName(properties[taskProperties.taskName]);
         task.changeCompleteness(properties[taskProperties.taskCompleted]);
-        if (rewriteAttachment) {
+
+        if (properties[Task.getTaskShouldUpdateAttachmentPropertyName()] === true) {
             let result;
             if (attachment) {
                 result = Utils.rewriteFolderWithAttachment(this.getAttachmentsDirectory(), String(taskId), attachment);
             } else {
                 Utils.deleteFolderWithAttachment(path.join(this.getAttachmentsDirectory(), String(taskId)));
             }
-            task.changeAttachment(result ? result.attachmentPath : null, result ? result.attachmentName : "No file");
+            task.changeAttachment(
+                result ? result.attachmentPath : null,
+                result ? result.attachmentName : "No file");
         }
+
+        return statuses.successNoContent;
     }
-    createTask(properties, attachment, rewriteAttachment, taskId) {
-        const createResult = Utils.rewriteFolderWithAttachment(this.getAttachmentsDirectory(), taskId, attachment);
+    createTask(properties, attachment) {
+
+        if (!Task.isValidObject(properties, false)) {
+            return statuses.unprocessableEntity;
+        }
+
+        const taskId = this.getNewItemIndex();
+        const createResult =
+            Utils.rewriteFolderWithAttachment(
+                this.getAttachmentsDirectory(),
+                String(taskId),
+                attachment);
 
         this.tasks[taskId] = new Task(
             taskId,
@@ -119,29 +187,8 @@ class TaskWorker {
                 : createResult.attachmentName,
             properties[taskProperties.taskCompleted]
         );
-    }
 
-
-    // properties is { },
-    // following properties of this object are defined:
-    // properties[taskProperties.taskDate]
-    // properties[taskProperties.taskCompleted]
-    // properties[taskProperties.taskName]
-
-    // attachment is undefined or file
-    insertTask(properties, attachment, rewriteAttachment, id = NEW_ITEM_INDEX) {
-
-        if (!this.isIndexValid(id)) {
-            throw new Error("invalid index passed");
-        }
-        const isTaskNew = id === NEW_ITEM_INDEX;
-        const taskId = isTaskNew ? this.getNewItemIndex() : id;
-
-        if (isTaskNew)  {
-            this.createTask(properties, attachment, rewriteAttachment, taskId);
-        } else {
-            this.changeTask(properties, attachment, rewriteAttachment, taskId);
-        }
+        return statuses.successNoContent;
     }
 }
